requirements.txt
Flask==2.0.1
Jinja2==3.0.3
Werkzeug==2.0.3
SQLAlchemy==1.4.35
Flask-SQLAlchemy==2.5.1
MarkupSafe==2.0.1
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-WTF==0.14.3
WTForms==2.3.3
email-validator==1.1.2
python-dotenv==0.17.1
itsdangerous==2.0.1
Pillow==10.2.0
psycopg2-binary==2.9.9
